Extract CreateResourceGroupStatement to reduce duplicate logic for MySQL and doris
